+++ lib/src/models/parameters/custom_version_compare_parameter_model.dart
@@ -1,11 +1,11 @@
-import '../../services/abstracts/json_to_version_response_service.dart';
+import '../../services/abstracts/version_convert_service.dart';
 import '../entities/entity_model.dart';
 import '../entities/store/base_store_model.dart';
 
 class CustomVersionCompareParameterModel<TData extends EntityModel<TData>> {
   final TData parseModel;
   final String currentAppVersion;
-  final JsonToVersionResponseService jsonToResponseService;
+  final VersionConvertService jsonToResponseService;
   final BaseStoreModel store;
   final String? Function(TData parseModel)? updateLinkGetter;
