+++ lib/src/services/abstracts/version_compare_service.dart
@@ -6,13 +6,13 @@ import '../../utils/constants/constants.dart';
 import '../../utils/mixins/launch_url_mixin.dart';
 import '../../utils/mixins/package_info_mixin.dart';
 import '../../utils/results/data_result.dart';
-import 'json_to_version_response_service.dart';
+import 'version_convert_service.dart';
 import 'remote_data_service.dart';
 
 abstract class VersionCompareService with PackageInfoMixin, LaunchUrlMixin {
   BaseStoreModel get store;
   RemoteDataService get dataService;
-  JsonToVersionResponseService get jsonToResponseService;
+  VersionConvertService get jsonToResponseService;
 
   Future<DataResult<VersionResponseModel>> getVersion();
 }
