+++ lib/src/services/abstracts/version_convert_service.dart
@@ -1,6 +1,6 @@
 import '../../models/entities/entity_model.dart';
 import '../../utils/results/data_result.dart';
 
-abstract class JsonToVersionResponseService<TData extends EntityModel<TData>> {
+abstract class VersionConvertService<TData extends EntityModel<TData>> {
   DataResult<String> convert(TData entity);
 }
