+++ lib/src/services/concretes/android/android_version_compare_manager.dart
@@ -1,13 +1,13 @@
-import '../../models/entities/empty_entity_model.dart';
-import '../../models/entities/store/android_store_model.dart';
-import '../../models/entities/store/base_store_model.dart';
-import '../../models/version_response_model.dart';
-import '../../utils/results/data_result.dart';
-import '../abstracts/json_to_version_response_service.dart';
-import '../abstracts/remote_data_service.dart';
-import '../abstracts/version_compare_service.dart';
-import 'android_json_to_version_response_manager.dart';
-import 'http_remote_data_manager.dart';
+import '../../../models/entities/empty_entity_model.dart';
+import '../../../models/entities/store/android_store_model.dart';
+import '../../../models/entities/store/base_store_model.dart';
+import '../../../models/version_response_model.dart';
+import '../../../utils/results/data_result.dart';
+import '../../abstracts/version_convert_service.dart';
+import '../../abstracts/remote_data_service.dart';
+import '../../abstracts/version_compare_service.dart';
+import 'android_version_convert_manager.dart';
+import '../http_remote_data_manager.dart';
 
 class AndroidVersionCompareManager extends VersionCompareByQueryService {
   AndroidVersionCompareManager({
@@ -17,11 +17,11 @@ class AndroidVersionCompareManager extends VersionCompareByQueryService {
   }) : store = AndroidStoreModel(bundleId);
 
   AndroidVersionCompareManager.httpService({
-    JsonToVersionResponseService? jsonToVersionResponseService,
+    VersionConvertService? jsonToVersionResponseService,
     required String bundleId,
   })  : store = AndroidStoreModel(bundleId),
         dataService = HttpRemoteDataManager(),
-        jsonToResponseService = jsonToVersionResponseService ?? AndroidJsonToVersionResponseManager();
+        jsonToResponseService = jsonToVersionResponseService ?? AndroidVersionConvertManager();
 
   @override
   final BaseStoreModel store;
@@ -30,7 +30,7 @@ class AndroidVersionCompareManager extends VersionCompareByQueryService {
   final RemoteDataService dataService;
 
   @override
-  final JsonToVersionResponseService jsonToResponseService;
+  final VersionConvertService jsonToResponseService;
 
   @override
   Future<DataResult<VersionResponseModel>> getVersion() async {
