+++ lib/src/services/concretes/custom/custom_version_compare_manager.dart
@@ -1,11 +1,11 @@
-import '../../models/entities/entity_model.dart';
-import '../../models/entities/store/base_store_model.dart';
-import '../../models/version_response_model.dart';
-import '../../utils/results/data_result.dart';
-import '../abstracts/json_to_version_response_service.dart';
-import '../abstracts/remote_data_service.dart';
-import '../abstracts/version_compare_service.dart';
-import 'http_remote_data_manager.dart';
+import '../../../models/entities/entity_model.dart';
+import '../../../models/entities/store/base_store_model.dart';
+import '../../../models/version_response_model.dart';
+import '../../../utils/results/data_result.dart';
+import '../../abstracts/version_convert_service.dart';
+import '../../abstracts/remote_data_service.dart';
+import '../../abstracts/version_compare_service.dart';
+import '../http_remote_data_manager.dart';
 
 class CustomVersionCompareManager<TData extends EntityModel<TData>> extends VersionCompareByQueryService {
   CustomVersionCompareManager({
@@ -32,7 +32,7 @@ class CustomVersionCompareManager<TData extends EntityModel<TData>> extends Vers
   final RemoteDataService dataService;
 
   @override
-  final JsonToVersionResponseService jsonToResponseService;
+  final VersionConvertService jsonToResponseService;
 
   final String currentAppVersion;
   final TData parseModel;
