+++ lib/src/services/concretes/huawei/huawei_version_compare_manager.dart
@@ -1,13 +1,13 @@
-import '../../models/entities/huawei/huawei_version_entity_model.dart';
-import '../../models/entities/store/base_store_model.dart';
-import '../../models/entities/store/huawei_store_model.dart';
-import '../../models/version_response_model.dart';
-import '../../utils/results/data_result.dart';
-import '../abstracts/json_to_version_response_service.dart';
-import '../abstracts/remote_data_service.dart';
-import '../abstracts/version_compare_service.dart';
-import 'http_remote_data_manager.dart';
-import 'huawei_json_to_version_response_manager.dart';
+import '../../../models/entities/huawei/huawei_version_entity_model.dart';
+import '../../../models/entities/store/base_store_model.dart';
+import '../../../models/entities/store/huawei_store_model.dart';
+import '../../../models/version_response_model.dart';
+import '../../../utils/results/data_result.dart';
+import '../../abstracts/version_convert_service.dart';
+import '../../abstracts/remote_data_service.dart';
+import '../../abstracts/version_compare_service.dart';
+import '../http_remote_data_manager.dart';
+import 'huawei_version_convert_manager.dart';
 
 class HuaweiVersionCompareManager extends VersionCompareByQueryService {
   HuaweiVersionCompareManager({
@@ -17,11 +17,11 @@ class HuaweiVersionCompareManager extends VersionCompareByQueryService {
   }) : store = HuaweiStoreModel(appId);
 
   HuaweiVersionCompareManager.httpService({
-    JsonToVersionResponseService? jsonToVersionResponseService,
+    VersionConvertService? jsonToVersionResponseService,
     required String appId,
   })  : store = HuaweiStoreModel(appId),
         dataService = HttpRemoteDataManager(),
-        jsonToResponseService = jsonToVersionResponseService ?? HuaweiJsonToVersionResponseManager();
+        jsonToResponseService = jsonToVersionResponseService ?? HuaweiVersionConvertManager();
 
   @override
   final BaseStoreModel store;
@@ -30,7 +30,7 @@ class HuaweiVersionCompareManager extends VersionCompareByQueryService {
   final RemoteDataService dataService;
 
   @override
-  final JsonToVersionResponseService jsonToResponseService;
+  final VersionConvertService jsonToResponseService;
 
   @override
   Future<DataResult<VersionResponseModel>> getVersion() async {
